+++ packages/hooks/test/create-hook-app.test.js
@@ -1,10 +1,43 @@
 const { resetState } = require('../src/state');
 const forrestjs = require('../src/index');
 const { registerAction } = require('../src/register-action');
+const {
+  isDeclarativeAction,
+  isListOfDeclarativeActions,
+} = require('../src/create-hook-app');
 const constants = require('../src/constants');
 
 describe('hooks/create-hook-app', () => {
   beforeEach(resetState);
+
+  describe('utils', () => {
+    const featureA = [
+      {
+        hook: 'foo',
+        handler: () => {},
+      },
+      {
+        hook: 'foo',
+        handler: [],
+      },
+      {
+        hook: 'foo',
+        handler: {},
+      },
+    ];
+
+    describe('isDeclarativeAction', () => {
+      featureA.forEach((payload) =>
+        it('Should recognise a declarative action payload', () => {
+          expect(isDeclarativeAction(payload)).toBe(true);
+        }),
+      );
+    });
+
+    describe('isListOfDeclarativeActions', () => {
+      expect(isListOfDeclarativeActions(featureA)).toBe(true);
+    });
+  });
 
   it('should run an empty app', async () => {
     await forrestjs.run();
@@ -39,7 +72,8 @@ describe('hooks/create-hook-app', () => {
     expect(handler.mock.calls.length).toBe(2);
   });
 
-  it('should register a service as single hook setup', async () => {
+  // DEPRECATED
+  it.skip('should register a service as single hook setup', async () => {
     const handler = jest.fn();
     const s1 = ['foo', handler];
     const f1 = ({ createHook }) => createHook('foo');
@@ -53,44 +87,53 @@ describe('hooks/create-hook-app', () => {
   });
 
   it('should run an app that provides settings as a function', async () => {
+    const f1 = jest.fn();
     await forrestjs.run({
       settings: ({ setConfig }) => {
         setConfig('foo.faa', 22);
       },
       features: [
-        [
-          constants.START_FEATURE,
-          ({ getConfig }) => expect(getConfig('foo.faa')).toBe(22),
-        ],
+        {
+          hook: '$START_FEATURE',
+          handler: ({ getConfig }) => f1(getConfig('foo.faa')),
+        },
       ],
     });
+
+    expect(f1.mock.calls[0][0]).toBe(22);
   });
 
   it('should provide a config getter to any registered action', async () => {
+    const f1 = jest.fn();
     await forrestjs.run({
       settings: ({ setConfig }) => {
         setConfig('foo.faa', 22);
       },
       services: [
-        // register a feature
+        // register a programmatic feature
         ({ registerAction }) =>
           registerAction({
-            hook: constants.INIT_SERVICE,
+            hook: '$INIT_SERVICE',
             handler: ({ getConfig, setConfig }) =>
               setConfig('foo', getConfig('foo.faa') * 2),
           }),
       ],
       features: [
-        // register a single action
-        [
-          constants.START_FEATURE,
-          ({ getConfig }) => expect(getConfig('foo')).toBe(44),
-        ],
+        // register a declarative feature
+        {
+          hook: '$START_FEATURE',
+          handler: ({ getConfig }) => f1(getConfig('foo')),
+        },
       ],
     });
+
+    expect(f1.mock.calls[0][0]).toBe(44);
   });
 
   it('should lock a context and decorate it with internal methods', async () => {
+    const f1 = jest.fn();
+    const f2 = jest.fn();
+    const f3 = jest.fn();
     await forrestjs.run({
       settings: {
         increment: 1,
@@ -99,26 +142,30 @@ describe('hooks/create-hook-app', () => {
         foo: (args, ctx) => args.value + ctx.getConfig('increment'),
       },
       services: [
-        [
-          constants.START_SERVICE,
-          async ({ createHook }) => {
+        {
+          hook: '$START_SERVICE',
+          handler: async ({ createHook }) => {
             const r1 = createHook.sync('aaa', { value: 1 });
-            expect(r1[0][0]).toBe(2);
+            f1(r1[0][0]);
 
-            const r2 = await createHook.serie('bbb', { value: 1 });
-            expect(r2[0][0]).toBe(2);
+            const r2 = await createHook.serie('bbb', { value: 2 });
+            f2(r2[0][0]);
 
-            const r3 = await createHook.parallel('ccc', { value: 1 });
-            expect(r3[0][0]).toBe(2);
+            const r3 = await createHook.parallel('ccc', { value: 3 });
+            f3(r3[0][0]);
           },
-        ],
+        },
       ],
       features: [
-        ['aaa', (args, ctx) => ctx.foo(args, ctx)],
-        ['bbb', (args, ctx) => ctx.foo(args, ctx)],
-        ['ccc', (args, ctx) => ctx.foo(args, ctx)],
+        { hook: 'aaa', handler: (args, ctx) => ctx.foo(args, ctx) },
+        { hook: 'bbb', handler: (args, ctx) => ctx.foo(args, ctx) },
+        { hook: 'ccc', handler: (args, ctx) => ctx.foo(args, ctx) },
       ],
     });
+
+    expect(f1.mock.calls[0][0]).toBe(2);
+    expect(f2.mock.calls[0][0]).toBe(3);
+    expect(f3.mock.calls[0][0]).toBe(4);
   });
 
   describe('createHookApp getters / setters', () => {
@@ -157,7 +204,7 @@ describe('hooks/create-hook-app', () => {
 
     it('should run a required service by reference', async () => {
       const handler = jest.fn();
-      const f1 = ['$S1', handler];
+      const f1 = { hook: '$S1', handler };
 
       await forrestjs.run({
         services: [s1],
@@ -169,7 +216,7 @@ describe('hooks/create-hook-app', () => {
 
     it('should fail to run a required service by reference', async () => {
       const handler = jest.fn();
-      const f1 = ['$S1', handler];
+      const f1 = { hook: '$S1', handler };
 
       let error = null;
       try {
@@ -186,7 +233,7 @@ describe('hooks/create-hook-app', () => {
 
     it('should ignore an optional service by reference', async () => {
       const handler = jest.fn();
-      const f1 = ['$S1?', handler];
+      const f1 = { hook: '$S1?', handler };
 
       await forrestjs.run({
         // services: [s1],
@@ -218,6 +265,42 @@ describe('hooks/create-hook-app', () => {
 
       expect(s1Handler.mock.calls.length).toBe(1);
       expect(s2Handler.mock.calls.length).toBe(1);
+    });
+  });
+
+  describe('run declarative features', () => {
+    it('Services and Features should be able to register a single service in a declarative way', async () => {
+      const handler1 = jest.fn();
+
+      await forrestjs.run([
+        {
+          hook: '$INIT_SERVICE',
+          handler: handler1,
+        },
+      ]);
+
+      expect(handler1.mock.calls).toHaveLength(1);
+    });
+
+    it('Services and Features should register multiple hooks in a declarative way', async () => {
+      const handler1 = jest.fn();
+      const handler2 = jest.fn();
+
+      await forrestjs.run([
+        [
+          {
+            hook: '$INIT_SERVICE',
+            handler: handler1,
+          },
+          {
+            hook: '$INIT_SERVICE',
+            handler: handler2,
+          },
+        ],
+      ]);
+
+      expect(handler1.mock.calls).toHaveLength(1);
+      expect(handler2.mock.calls).toHaveLength(1);
     });
   });
 });